Extract aes operations and symmetric crypto key to crypto module
