+++ crates/bitwarden/src/client/encryption_settings.rs
@@ -1,19 +1,11 @@
-use std::{collections::HashMap, str::FromStr};
+use std::collections::HashMap;
 
-use aes::cipher::{
-    block_padding::Pkcs7, generic_array::GenericArray, typenum::U32, BlockDecryptMut,
-    BlockEncryptMut, KeyIvInit,
-};
-use base64::Engine;
-use hmac::Mac;
-use rand::RngCore;
 use rsa::RsaPrivateKey;
 use uuid::Uuid;
 
 use crate::{
-    crypto::{CipherString, PbkdfSha256Hmac, PBKDF_SHA256_HMAC_OUT_SIZE},
+    crypto::{decrypt, decrypt_aes256, encrypt_aes256, CipherString, SymmetricCryptoKey},
     error::{CryptoError, Error, Result},
-    util::BASE64_ENGINE,
 };
 
 #[cfg(feature = "internal")]
@@ -22,76 +14,12 @@ use {
     rsa::{pkcs8::DecodePrivateKey, Oaep},
 };
 
-pub struct SymmetricCryptoKey {
-    pub key: GenericArray<u8, U32>,
-    pub mac_key: Option<GenericArray<u8, U32>>,
-}
-
-impl SymmetricCryptoKey {
-    const KEY_LEN: usize = 32;
-    const MAC_LEN: usize = 32;
-
-    pub fn generate(name: &str) -> Self {
-        use rand::Rng;
-        let secret: [u8; 16] = rand::thread_rng().gen();
-        crate::crypto::stretch_key(secret, name, None)
-    }
-
-    pub fn to_base64(&self) -> String {
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&self.key);
-
-        if let Some(mac) = self.mac_key {
-            buf.extend_from_slice(&mac);
-        }
-
-        BASE64_ENGINE.encode(&buf)
-    }
-}
-
-impl FromStr for SymmetricCryptoKey {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let bytes = BASE64_ENGINE
-            .decode(s)
-            .map_err(|_| CryptoError::InvalidKey)?;
-        SymmetricCryptoKey::try_from(bytes.as_slice())
-    }
-}
-
-impl TryFrom<&[u8]> for SymmetricCryptoKey {
-    type Error = Error;
-
-    fn try_from(value: &[u8]) -> Result<Self, Self::Error> {
-        if value.len() == Self::KEY_LEN + Self::MAC_LEN {
-            Ok(SymmetricCryptoKey {
-                key: GenericArray::clone_from_slice(&value[..Self::KEY_LEN]),
-                mac_key: Some(GenericArray::clone_from_slice(&value[Self::KEY_LEN..])),
-            })
-        } else if value.len() == Self::KEY_LEN {
-            Ok(SymmetricCryptoKey {
-                key: GenericArray::clone_from_slice(value),
-                mac_key: None,
-            })
-        } else {
-            Err(CryptoError::InvalidKeyLen.into())
-        }
-    }
-}
-
 pub struct EncryptionSettings {
     user_key: SymmetricCryptoKey,
     private_key: Option<RsaPrivateKey>,
     org_keys: HashMap<Uuid, SymmetricCryptoKey>,
 }
 
-// We manually implement these to make sure we don't print any sensitive data
-impl std::fmt::Debug for SymmetricCryptoKey {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Key").finish()
-    }
-}
 impl std::fmt::Debug for EncryptionSettings {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("EncryptionSettings").finish()
@@ -107,6 +35,7 @@ impl EncryptionSettings {
         private_key: CipherString,
     ) -> Result<Self> {
         // Stretch keys from the provided password
+
         let (key, mac_key) = crate::crypto::stretch_key_password(
             password.as_bytes(),
             auth.email.as_bytes(),
@@ -195,79 +124,6 @@ impl EncryptionSettings {
         let dec = encrypt_aes256(data, key.mac_key, key.key)?;
         Ok(dec)
     }
-}
-
-pub fn decrypt(cipher: &CipherString, key: &SymmetricCryptoKey) -> Result<Vec<u8>> {
-    match cipher {
-        CipherString::AesCbc256_HmacSha256_B64 { iv, mac, data } => {
-            let dec = decrypt_aes256(iv, mac, data.clone(), key.mac_key, key.key)?;
-            Ok(dec)
-        }
-        _ => Err(CryptoError::InvalidKey.into()),
-    }
-}
-
-pub fn decrypt_aes256(
-    iv: &[u8; 16],
-    mac: &[u8; 32],
-    data: Vec<u8>,
-    mac_key: Option<GenericArray<u8, U32>>,
-    key: GenericArray<u8, U32>,
-) -> Result<Vec<u8>> {
-    let mac_key = match mac_key {
-        Some(k) => k,
-        None => return Err(CryptoError::InvalidMac.into()),
-    };
-
-    // Validate HMAC
-    let res = validate_mac(&mac_key, iv, &data)?;
-    if res != *mac {
-        return Err(CryptoError::InvalidMac.into());
-    }
-
-    // Decrypt data
-    let iv = GenericArray::from_slice(iv);
-    let mut data = data;
-    let decrypted_key_slice = cbc::Decryptor::<aes::Aes256>::new(&key, iv)
-        .decrypt_padded_mut::<Pkcs7>(&mut data)
-        .map_err(|_| CryptoError::KeyDecrypt)?;
-
-    //Data is decrypted in place and returns a subslice of the original Vec, to avoid cloning it, we truncate to the subslice length
-    let decrypted_len = decrypted_key_slice.len();
-    data.truncate(decrypted_len);
-
-    Ok(data)
-}
-
-pub fn encrypt_aes256(
-    data_dec: &[u8],
-    mac_key: Option<GenericArray<u8, U32>>,
-    key: GenericArray<u8, U32>,
-) -> Result<CipherString> {
-    let mac_key = match mac_key {
-        Some(k) => k,
-        None => return Err(CryptoError::InvalidMac.into()),
-    };
-
-    let mut iv = [0u8; 16];
-    rand::thread_rng().fill_bytes(&mut iv);
-    let data = cbc::Encryptor::<aes::Aes256>::new(&key, &iv.into())
-        .encrypt_padded_vec_mut::<Pkcs7>(data_dec);
-
-    let mac = validate_mac(&mac_key, &iv, &data)?;
-
-    Ok(CipherString::AesCbc256_HmacSha256_B64 { iv, mac, data })
-}
-
-fn validate_mac(mac_key: &[u8], iv: &[u8], data: &[u8]) -> Result<[u8; 32]> {
-    let mut hmac = PbkdfSha256Hmac::new_from_slice(mac_key).expect("HMAC can take key of any size");
-    hmac.update(iv);
-    hmac.update(data);
-    let mac: [u8; PBKDF_SHA256_HMAC_OUT_SIZE] = (*hmac.finalize().into_bytes())
-        .try_into()
-        .map_err(|_| CryptoError::InvalidMac)?;
-
-    Ok(mac)
 }
 
 #[cfg(test)]
