+++ crates/bitwarden/src/crypto/mod.rs
@@ -1,9 +1,8 @@
 //! Cryptographic primitives used in the SDK
 
+use crate::error::CryptoError;
 use aes::cipher::{generic_array::GenericArray, typenum::U64, Unsigned};
 use hmac::digest::OutputSizeUser;
-
-use crate::client::encryption_settings::SymmetricCryptoKey;
 
 #[cfg(feature = "internal")]
 use aes::cipher::typenum::U32;
@@ -18,6 +17,10 @@ mod cipher_string;
 pub use cipher_string::CipherString;
 mod encryptable;
 pub use encryptable::{Decryptable, Encryptable};
+mod aes_opt;
+pub use aes_opt::{decrypt_aes256, encrypt_aes256};
+mod symmetric_crypto_key;
+pub use symmetric_crypto_key::SymmetricCryptoKey;
 
 #[cfg(feature = "internal")]
 mod fingerprint;
@@ -110,6 +113,16 @@ pub(crate) fn stretch_key(secret: [u8; 16], name: &str, info: Option<&str>) -> S
     hkdf.expand(i, &mut key).unwrap();
 
     SymmetricCryptoKey::try_from(key.as_slice()).unwrap()
+}
+
+pub fn decrypt(cipher: &CipherString, key: &SymmetricCryptoKey) -> Result<Vec<u8>> {
+    match cipher {
+        CipherString::AesCbc256_HmacSha256_B64 { iv, mac, data } => {
+            let dec = decrypt_aes256(iv, mac, data.clone(), key.mac_key, key.key)?;
+            Ok(dec)
+        }
+        _ => Err(CryptoError::InvalidKey.into()),
+    }
 }
 
 #[cfg(test)]
